requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
passlib
bcrypt
pyjwt
sqlalchemy
psycopg2-binary
python-multipart
httpx
schedule
